requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
django==4.1
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2; python_version >= '3.5'
tzdata==2022.2; sys_platform == 'win32'
